Create new PerfMonitor utility
